% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{copy_git_obj}
\alias{copy_git_obj}
\title{Copy the contents of a git_blob object to disk}
\usage{
copy_git_obj(obj, filename = sha(obj), path = ".")
}
\value{
The path to the copied file, which is named for the SHA of the blob
}
\description{
Copy the contents of a git_blob object to disk
}
