% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{copy_git_tree}
\alias{copy_git_tree}
\title{Copy the contents of a git_tree object to a directory on disk recursively}
\usage{
copy_git_tree(tree, dirname = sha(tree), path = ".")
}
\value{
The path to the directory, which is named for the SHA of the tree
}
\description{
Copy the contents of a git_tree object to a directory on disk recursively
}
