% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_repo.R
\name{create_example_repo}
\alias{create_example_repo}
\title{Set up a testing and example repository}
\usage{
create_example_repo(
  dir = fs::file_temp("relic_example_"),
  reporter = "silent",
  s3 = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{dir}{Path to the directory where the repository should be created}

\item{reporter}{the reporter to use when building targets with \code{\link[targets:tar_make]{targets::tar_make()}}. Defaults to "silent".}

\item{s3}{Whether the repository should use S3 storage for targets. Note that
the S3 endpoint and bucket must already be available.}

\item{overwrite}{Whether to overwrite the directory if it already exists.}
}
\value{
The path to the created repository
}
\description{
This function generates a repository with a commit history that can be used
for testing and examples.
}
