% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy_version.R
\name{file_copy_version}
\alias{file_copy_version}
\alias{file_copy_versions}
\title{Copy files from git history}
\usage{
file_copy_version(
  path,
  ref = "HEAD",
  name = NULL,
  dir = NULL,
  full_name = NULL,
  recurse = TRUE,
  repo = ".",
  use_cache = TRUE
)

file_copy_versions(
  path,
  from = "HEAD",
  to = NULL,
  name = NULL,
  dir = NULL,
  full_name = NULL,
  recurse = TRUE,
  repo = ".",
  use_cache = TRUE
)
}
\arguments{
\item{path}{A path or vector of paths of files and/or folders to extract,
relative to the git directory}

\item{ref}{A git commit SHA, tag, branch or other \href{https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions}{revision string}:
such as "HEAD~1", or a \code{\link[git2r:commit]{git2r::commit()}} object. Defaults to "HEAD".}

\item{name}{the name to give the file when it is copied to disk. Defaults to
the original name of the file or directory.}

\item{dir}{The path to the directory where the file should be copied.
Defaults to a system-level temporary directory named after the SHA of the
git commit. Will be created if does not exist.}

\item{full_name}{If TRUE, the file will be written to the full relative path
of its location in the git repository, below \code{dir}. Defaults to TRUE if
both \code{name} and \code{dir} are NULL, FALSE otherwise.}

\item{recurse}{If \code{path} is a directory, should should it be copied recursively? If
numeric, how many levels deep should it be copied? Defaults to TRUE.}

\item{repo}{The path to the git repository or a \code{\link[git2r:repository]{git2r::repository()}} object.}

\item{use_cache}{Should the cache be used? Defaults to TRUE.}

\item{from}{A commit object or revision string. A string of the form
'from...to' may also be passed, in which case the commits between the two
revision strings are used and the \code{to} argument is ignored. If a list of commits
is passed, these commits are used rather than calculating the commits between.}

\item{to}{A commit object or reference}
}
\value{
\itemize{
\item For \code{file_copy_version()}, A character vector of paths to the copied files or directories.
\item For \code{file_copy_versions()}, a list of character vectors of paths.
Where files do not exist in a commit, \code{NA} values are returned.
}
}
\description{
Copy files from git history to disk. Directories are copied recursively.
}
\details{
The default behavior is to copy the file to a cache directory named after
both the commit and and the file path. These paths are universally unique,
so if \code{use_cache = TRUE}, the file will not be re-copied if it already exists
at that path.
}
