% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relic.R
\name{is_relic}
\alias{is_relic}
\title{An internal S3 class of git objects with additional metadata}
\usage{
is_relic(x)
}
\arguments{
\item{x}{an object}
}
\description{
An internal S3 class of git objects with additional metadata
}
