% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relic-package.R
\docType{package}
\name{relic-package}
\alias{relic}
\alias{relic-package}
\title{Objects from a while back}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'relic' package provides tools for extracting files and objects from the history of a git repository. It is a high-level interface designed to enable comparison of objects in reproducible research workflows, especially pipelines that use the 'targets' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ecohealthalliance.github.io/relic}
  \item \url{https://ecohealthalliance.r-universe.dev/relic}
  \item Report bugs at \url{https://github.com/ecohealthalliance/relic/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID})

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder, funder]
}

}
\keyword{internal}
