% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relic_cache}
\alias{relic_cache}
\alias{relic_cache_clear}
\title{The relic cache directory}
\usage{
relic_cache()

relic_cache_clear()
}
\value{
The path to the relic cache directory
}
\description{
Get the relic cache directory, which can be specified as either the R option
\code{relic.cache.dir} or the environment variable \code{RELIC_CACHE_DIR}. (The
environment variable has higher priority).  If neither
is set, the default is set by tools::R_user_dir("relic", "cache").
}
\examples{
relic_cache()
}
