% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_exists_version.R
\name{tar_exists_version}
\alias{tar_exists_version}
\alias{tar_exists_version_raw}
\title{Check if a target exists in a version of the pipeline}
\usage{
tar_exists_version(
  name,
  ref = "HEAD",
  branches = NULL,
  repo = ".",
  store = targets::tar_path_store()
)

tar_exists_version_raw(
  name,
  ref = "HEAD",
  branches = NULL,
  repo = ".",
  store = targets::tar_path_store()
)
}
\arguments{
\item{name}{Name of the target. \code{tar_read_version()} can take a symbol, \code{tar_read_raw_version()} requires a character.}

\item{ref}{A git commit SHA, tag, branch or other \href{https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions}{revision string}:
such as "HEAD~1", or a \code{\link[git2r:commit]{git2r::commit()}} object. Defaults to "HEAD".}

\item{branches}{Integer of indices of the (targets) branches to load if the target is a pattern.}

\item{repo}{The path to the git repository or a \code{\link[git2r:repository]{git2r::repository()}} object.}

\item{store}{Path to the targets store within the project. Defaults to the current project's store.}
}
\description{
Check if a target exists in a version of the pipeline
}
