% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_load_version}
\alias{tar_load_version}
\alias{tar_read_version}
\title{Get a target from a given git reference}
\usage{
tar_load_version(name, ref = "HEAD", envir = parent.frame(), ..., repo = ".")

tar_read_version(name, ref = "HEAD", repo = ".", ...)
}
\arguments{
\item{name}{The name of the target}

\item{ref}{The git reference: tag, branch, SHA, or revision like HEAD~1 (see https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions)}

\item{...}{Additional arguments passed to \code{\link{tar_load}} or  \code{\link{tar_read}}}
}
\description{
Get a target from a given git reference
}
