% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_meta_version.R
\name{tar_meta_version}
\alias{tar_meta_version}
\title{Read a target project's metadata from git history}
\usage{
tar_meta_version(
  ref = "HEAD",
  ...,
  store = targets::tar_path_store(),
  repo = "."
)
}
\arguments{
\item{ref}{A git commit SHA, tag, branch or other \href{https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions}{revision string}:
such as "HEAD~1", or a \code{\link[git2r:commit]{git2r::commit()}} object. Defaults to "HEAD".}

\item{...}{Arguments passed to \code{\link[targets:tar_meta]{targets::tar_meta()}}}

\item{store}{Path to the targets store within the project. Defaults to the current project's store.}

\item{repo}{The path to the git repository or a \code{\link[git2r:repository]{git2r::repository()}} object.}
}
\description{
Read a target project's metadata from git history
}
