% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{relic_cache}
\alias{relic_cache}
\alias{relic_cache_delete}
\alias{relic_cache_cleanup}
\alias{relic_cache_regular_cleanup}
\alias{relic_cache_max_size}
\alias{relic_cache_max_age}
\alias{relic_cache_cleanup_time}
\title{The relic cache}
\usage{
relic_cache()

relic_cache_delete()

relic_cache_cleanup(
  max_age = relic_cache_max_age(),
  max_size = relic_cache_max_size()
)

relic_cache_regular_cleanup(cleanup_time = relic_cache_cleanup_time())

relic_cache_max_size()

relic_cache_max_age()

relic_cache_cleanup_time()
}
\arguments{
\item{max_age}{The maximum age of files to keep in the cache, as a \link[base:difftime]{difftime}
object.  Files older than this will be deleted.  Defaults to \code{Inf}. Can be
set with the environment variable \code{RELIC_CACHE_MAX_AGE} or
\code{options("relic.cache.max.age")}, which take numeric time in days or a
string with units, e.g., "1 day" or "2 weeks".}

\item{max_size}{The maximum size of the cache, as a string that can be parsed
by \code{\link[fs:fs_bytes]{fs::fs_bytes()}}.  Defaults to "20 MB".  Can be set with the environment
variable \code{RELIC_CACHE_MAX_SIZE} or \code{options("relic.cache.max.size")}.
Cached files will be deleted from oldest to youngest until the cache size
is under this limit.}

\item{cleanup_time}{The time between cache cleanups, as a \link[base:difftime]{difftime} object.
Defaults to 1 day. Can be set with the environment variable
\code{RELIC_CACHE_CLEANUP_TIME} or \code{options("relic.cache.cleanup.time")}, which
take numeric time in days or a string with units, e.g., "1 day" or "2
weeks".  If set to "Inf", no cleanup will be performed at startup.}
}
\value{
The path to the relic cache directory
}
\description{
The relic cache directory stores files that have been retrieved
from both local and remote repositories to avoid repeated extractions or
downloads. Its location can be set with the environment variable
\code{RELIC_CACHE_DIR} or \code{options("relic.cache.dir")}, and it defaults to the
user cache directory.  The cache is cleaned up regularly at package startup,
but can also be cleaned up manually with \code{relic_cache_cleanup()} or cleared
entirely with \code{relic_cache_delete()}.
}
\examples{
relic_cache()
}
