% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_meta_version.R
\name{tar_meta_version}
\alias{tar_meta_version}
\title{Read a target project's metadata from repository history}
\usage{
tar_meta_version(
  ref = "HEAD",
  repo = ".",
  store = NULL,
  endpoint = NULL,
  region = NULL,
  ...
)
}
\arguments{
\item{ref}{A git commit SHA, tag, branch or other \href{https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions}{revision string}:
such as \code{"HEAD~1"}, \code{"main@{2023-02-26 18:30:00}"}, or
\code{"branch@{yesterday}"}. Can also be a \code{\link[git2r:commit]{git2r::commit()}} object.  For S3
buckets, a VersionID. Defaults to "HEAD", which also means the latest
version in an S3 bucket.}

\item{repo}{The repository to get the file from. This can be a local git
directory, a GitHub repository (URL or \code{"owner/repo"} string), or an S3
bucket indicated by "s3://bucket-name". Defaults to the current working
directory.}

\item{store}{Path to the targets store within the project. Defaults to
\verb{"_targets}, or the current project's \link[targets:tar_path_store]{store name} if \code{repo = "."}.}

\item{endpoint, region}{For S3 buckets, the endpoint and region of the bucket.}

\item{...}{Arguments passed to \code{\link[targets:tar_meta]{targets::tar_meta()}}}
}
\value{
A data frame with one row per target/object. See
\code{\link[targets:tar_meta]{targets::tar_meta()}} for details.
}
\description{
This function extracts \link[targets:tar_meta]{targets metadata} from versioned
history. In most cases this is a git repository, but it can also be an S3
cloud bucket for a project using cloud versioning and storing the metadata
file in the cloud. (See \code{repository_meta} in \code{\link[targets:tar_option_set]{targets::tar_option_set()}}).
}
